Adds Jasmine test infra.
This introduces a single Jasmine test just to verify that it is working. It also includes a new TypeScript project for compiling test code to a separate directory so they aren't shipped to NPM. I also added a solution-style tsconfig so Jasmine types are limited to test files.

Test files are simply built before execting Jasmine. This means there is no fancy live-reload system, but I don't think that is too important for a project of this size.