Adds examples/tests of type inference.
This verifies that the type system works as expected. This is mostly done by abusing TypeScript's `@ts-expect-error` annotation to force the compiler to fail a test build if one of these invariants do not hold.

This also includes tests of things which should work but actually don't. Since this is only a library and built on top of TypeScript's build system, there is only so much that can be done. This serves partially as documentation of the differences between this library and the actual paradigm.