Adds examples/tests of using factories with `ctor<T>`.
Factories are mostly required to do anything useful with this system so they are the most simplistic use case.

This includes a quick refresher of how factories are traditionally difficult to compose with each other. Using `ctor<T>` we can nicely abstract the return value of a superclass' factory in a usable and encapsulated fashion.