Adds examples/tests of dependency injection of `ctor<SuperClass>`.
Using normal constructors, if a superclass supports multiple inputs to its constructor, all its subclasses must also support the same, which can be quite awkward and involve information not relevant to the subclass. By using `ctor<T>`, users of a subclass can inject the `ctor<SuperClass>` object they wish to construct with, effectively deffering the responsibility of providing superclass constructor parameters to users of the subclass, rather than proxying them through the subclass constructor itself.

In TypeScript this is implemented via unions, but in other OO languages, this is often implemented via function overloading of constructors. The same problem exists with either model.