Adds examples/tests showing the coupling/decoupling behaviors of constructors.
In normal constructors, a superclass is constructed at the same time as the subclass. This limits the work that can be done between the two and makes constructor implementations awkward. This example shows how `ctor<T>` factories can easily perform complex or long-running operations at any time during construction in an intuitive fashion.