Adds examples/tests of an `OptimizedSet` use case.
This is a real-world use case which leverages dynamically choosing a superclass as part of its implementation. The desired behavior is a `createOptimizedSet()` function which dynamically chooses between two implementations of a simple `Set`. One is optimized for very small sets (`SparseSet`), while the other is optimized for very large sets (`DenseSet`). This is a trivial factory function on its own. However, the tricky part here is a `createMutableOptimizedSet()` factory which should reuse the existing functionality and optimization, but including the functionality to add items to the set afterwards.

There is no good way to implement a mutable extension of an `OptimizedSet` with traditional constructors, because there are two implementations of `OptimizedSet` which are decided between in the `createOptimizedSet()` factory. This factory is impractical to compose with a subclass. I saw two ways of implementing this that both kind of suck.

The first one is to have a `MutableSparseSet` which extends `SparseSet` and a separate `MutableDenseSet` which extends `DenseSet`. This requires the mutability feature to be duplicated between the two classes, with no easy way of sharing the implementation. It also means that the `createMutableOptimizedSet()` factory must copy the decision making logic of `createOptimizedSet()`, except reference the mutable versions of the classes.

The second attempt is to implement mutability as a traditional JavaScript/TypeScript mixin which can be applied to either implementation of `OptimizedSet`, and then making a generic `optimize()` function to encapsulate the optimization decision in a generic fashion. `MutableSparseSet` and `MutableDenseSet` are then empty classes which extended their immtuable equivalents with the `Mutable` mixin applied. The `optimize()` function works ok, although it is fiddly and complicates the implementation significantly. The mixin works initially, but falls apart as soon as you try to add a constructor parameter. TypeScript generally requires that a mixin constructor is implemented with `any[]` parameters, meaning `Mutable` cannot apply its own parameters. While there are tricks to work around this, they ultimately ran into other shortcomings of constructors. For example, you also cannot use field initializers if `super()` is not the first statement of a constructor. In the end, I decided that a mixin implementation with a constructor parameter was too difficult to work and just gave up. That honestly kind of proves my point that constructors are such a pain to deal with, so I don't feel too bad for failing to make this work. Even if I did, the additional constraints and complexity would greatly dimish the value of using the mixin pattern here.

Using `ctor<T>` works quite well here however, because it does not require a declaration-time reference to the superclass. `MutableSet` simply implements a unknown implementation of `OptimizedSet`, which could be satisfied by either a `SparseSet` or a `DenseSet`. It cleanly compose `createOptimizedSet()` and adds mutation functionality without requiring awkward mixins, duplicated code, or extending every implementation of `OptimizedSet`. This also allows constructor parameters to `MutableSet`, which is where the mixin implementation starts to fall over. This is a much cleaner implementation than non-`ctor<T>` alternatives.