Uses a manual prototype implementation instead of relying on the standard JS implementation.
In order to support mixins and dynamically choosing superclasses at construction-time, it is necessary to remove inheritence as a declaration-time binding. Rather than a class declaring what class it extends, it declares a type constraint and is provided a class which conforms to that constraint during construction. The JavaScript object model fundamentally doesn't work this way and required completely rearchitecting the way that object construction is performed to support this.

With this new model, all subclasses actually inherit directly from `Base` in a JavaScript context. This `Base` class doesn't really do anything, and it means that a subclass and superclass are actually unrelated as far as JavaScript is concerned.

To make this work as a user expects, `ctor<T>.prototype.construct()` is responsible for not just creating an object and invoking its constructor, but also invoking all parent class constructors and  setting up its object hierarchy. It manually assigns prototypes during construction, meaning that the inheritance heirarchy is specified at **construction-time**, not at **declaration-time**, making `ctor<T>` much more powerful than traditional constructors. This implementation relied on copying and mangling prototypes to build this new hierarchy during construction. It also means it broke the `instanceof` operator, so I had to manually override `Symbol.hasInstance` to fix it.

One other effect of this change is that the `super` keyword is no longer useful. In constructors, `super()` actually refers to the constructor of `Base`, rather than the actual superclass constructor (which is invoked by `ctor<T>.prototype.construct()`). This is because even transpiled to ES5, `super` is implemeted via a direct reference to the declared superclass. Since `Base` doesn't really do anything in its constructor, it means we can just call `super()` because it won't really do anything meaningful. This does have the added benefit of making constructors smaller and no longer relying on it to pass through superclass arguments. Similarly, `super.method()` is also broken because again, it refers to `Base`, which doesn't implement any methods. `Base` provides a `_super` property which emulates this behavior, returning the real superclass prototype with methods bound so `this._super.method()` works as expected.

Classes must now extend `Implementation<SuperClass>()` rather than extend `SuperClass` directly as they generally would. This just returns `Base`, but typed as if it is `SuperClass`, meaning TypeScript will believe the class extends `SuperClass`, but in reality, it will extend `Base`. `ctor<T>.prototype.construct()` is actually responsible for setting up the prototypes to make this work as a user would expect. `Implementation<T>()` is a bit of an awkward abstraction, but it means that a user can extend a class based only on its type, and not its implementation, which allows for some really cool mixin functionality.

For example, this means that superclass can be chosen in a factory's implementation. For example, a factory can dynamically choose, based on its inputs, which of multiple superclasses to inherit from. This simply relies on the fact that the inheritance hierarchy is not built until construction time, so one instance of a child class may extend `ParentA`, while a second instance may extend `ParentB`. As long as both parents satisfy the interface required by the child class, this is still totally valid and usable.

I also updated the basic tests to include method calls to verify that the prototype chain is working as expected. This also fixes static factories using the same name as a superclass factory, which is a nice bonus.

This unfortunately requires transpiling code to ES5, meaning we aren't actually using real `class` syntax at runtime. This is because `ctor<T>.prototype.construct()` must manually invoke the constructor, which is not allowed using `class` syntax without a `new` call. The easiest way around this limitation is to transpile to ES5, so everything uses simple `function` syntax, rather than `class`.