Adds examples/tests of mixins.
With recent changes to the way construction works, it is no longer necessary for a class to have a direct reference to its superclass during its declaration. In fact, extending `Implementation<T>()` only includes a type reference to the superclass, not a value reference. This means subclasses do not need a reference to their actual superclass! This is exactly what mixin are for, extending a class with only limited knowledge of it. I included some examples of mixing in to an unknown class, as well as mixing in to an unknown implementation of an interface.

Interfaces are often touted as the single-inheritence solution to many features provided by multiple inheritence. However, many developers often complain that interfaces are not sufficient as an stand-in for multiple inheritence, because they only provide limited polymorphism. They don't provide implementations to their methods and they cannot introduce additional state to an object. Mixins are a way of providing this functionality without breaking single-inheritence.

Using `ctor<T>`, mixins actually work quite well. All it takes to define one is to extend `Implementation()` and construct using a generic `ctor<Parent>` using `from(parentCtor).mixin()`, then return `ctor<Mixin & Parent>`. No complicated function-which-wraps-a-parent-class-into-a-child-class-with-the-mixin nonsense. `ctor<T>.prototype.mixin()` is basically equivalent to `ctor<T>.prototype.new()`, except that it does not require the subclass to extend the superclass. This allows mixins to be flexible enough to extend any parent class.

Mixins can also support constraints by extending `Implementation<SomeInterface>()`, meaning that contructing a `ctor<Mixin>` requires a `ctor<SomeInterface>`. I like to think of this as a mixin which "extends some implementation of this interface". The mixin knows the contract of the interface, but does not have any knowledge of the actual implementation of its superclass.

One unfortunate caveat is that `from(parentCtor).mixin(MyMixin)` cannot automatically verify that `parentCtor` actually provides the superclass of `MyMixin`. It is up to the mixin itself to define `parentCtor` as a generic type which implements the base interface required by `MyMixin`. This isn't that hard, but it can be easy to forget. Included this in the documentation and invariants list to draw some attention to this foot-gun.