Removes ES5 transpilation.
The core need for ES5 was that native `class` syntax requires construction to use the `new` operator. I reworked `.constructOn()` to use `new` and then simply copy the result over to the object intended as `this`. Using this strategy has the same effect but uses `new` in order to be compatible with native `class` syntax.

The one other side effect is that the `prototype` of native `class` syntax has all its properties marked as non-enumerable. This means they do not show up in `Object.entries()` or even `for-in`. `Object.getOwnPropertyNames()` is necessary to enumerate these values and copy them to the new prototype.

Also I previously only applied the ES5 target onto the test compilation, so this was already broken for real usage anyways. Whoops!