Adds test case for constructing a class off an incorrect subclass of the superclass.
This example shows two classes intending to be siblings of a single superclass. However, the type system allows one sibling to extend the other in an unintended and unexpected way. This could potentially lead to awkward bugs.

Fortunately, most factories of a concrete class would return an already constructed instance, rather than a `ctor<T>`. This means such classes cannot be extended anyways and would be protected from such a mistake. However, there are other uses for `ctor<T>` beyond inheritance, and so a developer mistake here is still possible.

I'm noting this down as a test case for now, however I think the solution would involve an `extendableCtor<T>` vs `ctor<T>`. Any class intended to be final should be explicitly protected from extension by a subclass.

Of course, if the subclass also happens to be extendable, this issue would still be possible, but there is no way to know whether or not a developer actually intended to have two classes as siblings or not, meaning it is impossible to fully protect devs from this potential foot-gun. Requiring the class to be non-final (with final being a default), may be protection enough for the majority of cases.