Adds examples/tests of a framework use case.
This simulates a simple framework which is responsible for managing and constructing user-defined classes. This is traditionally very awkward because constructor are so tricky to work with.

Using `ctor<T>`, we can nicely encapsulate the parameters for the framework-owned superclass and isolate them from the user-owned subclass.

Honestly I don't think this is the best example, as a factory function is required in either case in order to provide the 'World' string. It is still slightly better and I think it is an important use case to consider. Hopefully I can come up with another example which more has clearer benefits in the future.